third_party/blink/web_tests/platform/mac/external/wpt/event-timing/dblclick-expected.txt

This is a testharness.js-based test.
FAIL Event Timing dblclick. assert_greater_than: The entry's processingStart should be greater than startTime. expected a number greater than 2077.215000000251 but got 2077.215000000251
Harness: the test ran to completion.
